Missle Blast Radius
I have some code in for a blast radius on the missle although it doesn't
appear to be working properly. I tried adding a new bounding box to
check against and it made the particle system go crazy so i stuck with
comparing the position of the missle to the positions of the enemies but
its still not working. Any suggestions guys?